refactor: build
